Fixed InputStreamReader to use UTF-8
Per forecast.io API docs, the JSON responses are in UTF-8 format due to the text summaries that can contain unicode characters.

Therefore, updated the calls to create the InputStreamReader to specify UTF-8.

Otherwise, with unicode characters in the text summaries (such as the "degree" symbol) get replaced with question marks (?) in the JSON.